Move RedisUtils to utils.py
